Routes was in the wrong directory - moved it to lib as a sub directory and changed the include statements to match.
