+++ lib/app.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const express = require('express');
-const city = require('../routes/router');
+const city = require('./routes/router');
 const errorHandler = require('./error-handler');
 
 const app = express();
