+++ lib/routes/router.js
@@ -2,40 +2,25 @@
 
 const express = require('express');
 const router = express.Router();
-const ds = require('../lib/datastore');
+const path = require('path');
+const ds = require('../datastore');
+const bodyParser = require('../bodyHandler');
 const pathname = path.join(__dirname, '..', 'city');
 
 router
-  .post('/city/:id', (req, res, next) => {
+  .post('/city/:id', bodyParser, (req, res, next) => {
     console.log('Writing new file...');
-    let body = '';
-    req.on('data', data => body += data);
-    req.on('end', () => {
-      var filename = req.params.id;
-      if (filename.slice(-5) !== '.json') {
-        filename += '.json';
-      }
-      ds.write(pathname, filename, body)
-        .then(() => res.send(`${filename} saved.`))
-        .catch(next);
-    });
+    ds.write(pathname, req.filename, req.body)
+      .then(() => res.send(`${req.filename} saved.`))
+      .catch(next);
   })
-  .put('/city/:id', (req, res, next) => {
+  .put('/city/:id', bodyParser, (req, res, next) => {
     console.log(`Updating ${req.params.id} file.`);
-    let body = '';
-    req.on('data', data => body += data);
-    req.on('end', () => {
-      var filename = req.params.id;
-      if (filename.slice(-5) !== '.json') {
-        filename += '.json';
-      }
-      const filepath = path.join(pathname, filename);
-      ds.deleteOne(filepath)
-        .catch(next('fs delete error', err.code))
-        .then(() => { ds.write(pathname, filename, body); })
-        .then(() => { res.send( `File ${filename} saved.`); })
-        .catch(next);
-    });
+    const filepath = path.join(pathname, req.filename);
+    ds.deleteOne(filepath)
+      .then(() => { ds.write(pathname, req.filename, req.body); })
+      .then(() => { res.send( `File ${req.filename} saved.`); })
+      .catch(next);
   })
   .get('/city', (req, res, next) => {
     ds.getAll(pathname)
